fix(大量炒饭修复):高级战斗，皮肤

+++ function/core/Todo.py
@@ -983,8 +983,9 @@ class ThreadTodo(QThread):
             self.thread_card_manager = None
 
             if self.opt["senior_settings"]["auto_senior_settings"]:
-                kill_process(self.process)
-                self.process = None
+                if senior_setting:
+                    kill_process(self.process)
+                    self.process = None
 
             CUS_LOGGER.debug('thread_card_manager 退出事件循环并完成销毁线程')
 
@@ -1240,12 +1241,13 @@ class ThreadTodo(QThread):
             stage_plan_by_id = get_stage_plan_by_id()
             skip = stage_plan_by_id["skip"]
             deck = stage_plan_by_id["deck"]
-            try:
-                senior_setting = stage_plan_by_id["senior_setting"]
-            except KeyError:
-                pass
             battle_plan_1p = stage_plan_by_id["battle_plan"][0]
             battle_plan_2p = stage_plan_by_id["battle_plan"][1]
+        try:
+            stage_plan_by_id = get_stage_plan_by_id()
+            senior_setting = stage_plan_by_id["senior_setting"]
+        except KeyError:
+            pass
 
         faa_a = self.faa_dict[pid_a]
         faa_b = self.faa_dict[pid_b] if pid_b else None
